+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 ### New
 
 - TODO ([#XXX](https://github.com/kedacore/keda/pull/XXX))
+- ScaledJob: introduce MultipleScalersCalculation ([#2016](https://github.com/kedacore/keda/pull/2016))
+- Add Graphite Scaler ([#1628](https://github.com/kedacore/keda/pull/2092))
 
 ### Improvements